pythonfast/pyAgentDemo/pyproject.toml
[project]
name = "pyagentdemo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=3.1.1",
    "langchain>=1.2.13",
    "langchain-anthropic>=1.4.0",
    "langchain-core>=1.2.22",
    "langchain-mcp-adapters>=0.2.2",
    "langchain-openai>=1.1.12",
    "langgraph>=1.1.3",
    "python-dotenv>=1.2.2",
]

[project.scripts]
start = "pyagentdemo.main:main"

[tool.uv]
dev-dependencies = []
